requirements.txt
torch
torchvision
pillow
numpy
opencv-python
scikit-learn
tqdm
gdown
albumentations
mamba-ssm
numpy
timm
einops
fastprogress
pytorch-lightning
fvcore
tensorflow